Fixed Packet Sending
I made some changes so that the buffer gets copied correctly. With
this, the buffer passing for sending is complete. The whole thing
still has the `FOMDataUnion` for interop though, so next I need
to remove the union completely and just start working with the
raw structs and their related buffers.