Use Contiguous Send Packet Buffer
Rather than pinning each QueuePacket's buffer, it is better for us to pin a
contiguous block of memory. This avoids heap fragmentation from a
bunch of small buffers that the CLR cannot move. We now copy
the contents into a buffer and then use pointes into that instead
of the union struct.