Reworked Packet Sending
After what feels like a lifetime of debugging, I tracked a memory corruption
bug down to the `using var response` for `QueuePacket.Create()`. This is
done to prevent exceptions from leaking memory, however, I did not
fully understand C#'s "in" parameter keyword. What was happening
was that the array was being returned to the pool on returning
from "handle" because the transfer happened on a defensive
copy. This new structure eliminates those copies and fixes
the memory corruption.